Wrap by _options_to_str
